Update Javascript parser, add unit test, fix README typo
