+++ README.md
@@ -1,2 +1,2 @@
 # Read from the standard input
-Here is some example code you can use to read a series of strings from the standard imput.
+Here is some example code you can use to read a series of strings from the standard input.
