+++ javascript/package.json
@@ -4,8 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "parse": "node index.js"
+    "test": "tape **/*.test.js"
   },
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "devDependencies": {
+    "tape": "^4.9.1"
+  }
 }
